test_requirements.txt
pytest
pytest-cov
python-coveralls
sh
flask-testing
blinker